подравлены миграции бд

+++ database/migrations/2021_04_09_163745_create_comments_table.php
@@ -14,11 +14,11 @@ class CreateCommentsTable extends Migration
     public function up()
     {
         Schema::create('comments', function (Blueprint $table) {
-            $table->bigIncrements('id');
+            $table->increments('id');
             $table->string('text',150);
             $table->date('date');
-            $table->bigInteger('id_user');
-            $table->bigInteger('id_project');
+            $table->integer('id_user');
+            $table->integer('id_project');
             $table->foreign('id_user')->references('id')->on('users');
             $table->foreign('id_project')->references('id')->on('projects');
             $table->timestamps();
